+++ docs/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * For players:
   * Added friendly error in 64-bit mode when a mod is 32-bit only.
+  * Fixed console encoding issues on Linux/macOS.
   * Fixed some installer errors now show info header.
 
 * For mod authors: